// src/types/tmo.ts
export interface TMOCustomAttribute {
  attribute_code: string;
  value: string;
}
// ...
export interface TMOProductOption {
  product_sku: string;
  option_id: number;
  title: string;
  type: string;
  sort_order: number;
  is_require: boolean;
  price?: number;
  price_type?: string;
  max_characters?: number;
  values?: TMOProductOptionValue[];
}

export interface TMOProductOptionValue {
  title: string;
  sort_order: number;
  price: number;
  price_type: string;
  option_type_id: number;
}
// ...
export interface TMOCartItemProductOption {
  extension_attributes?: {
    custom_options?: TMOCartItemCustomOption[];
    configurable_item_options?: TMOCartItemConfigurableOption[];
  };
}

export interface TMOCartItemCustomOption {
  option_id: string;
  option_value: string;
}

export interface TMOCartItemConfigurableOption {
  option_id: string;
  option_value: number;
}
// ...
export interface MappedProduct {
  id: number;
  sku: string;
  name: string;
  artist_name: string;
  description: string;
  price: number;
  personal_credits?: number;
  commercial_credits?: number;
  category: string;
  categoryIds: string[]; // All category IDs this product belongs to
  url: string;
  images: string[];
  is_pack: boolean;
  pack_items?: MappedProduct[];
  options?: TMOProductOption[];
  custom_attributes: TMOCustomAttribute[];
}